Continue migrating functions into switch GUI
